+++ src/App.js
@@ -1,13 +1,18 @@
 //node_module
-import React, { Component } from 'react';
-import { BrowserRouter, Route } from 'react-router-dom';
-import { library } from '@fortawesome/fontawesome-svg-core';
-import { faBars, faTimes, faAngleLeft, faAngleRight } from '@fortawesome/free-solid-svg-icons';
+import React, { Component } from "react";
+import { BrowserRouter, Route } from "react-router-dom";
+import { library } from "@fortawesome/fontawesome-svg-core";
+import {
+  faBars,
+  faTimes,
+  faAngleLeft,
+  faAngleRight
+} from "@fortawesome/free-solid-svg-icons";
 
 //container&components
-import { Login, ClassesList, Class, Exam } from 'container';
+import { Login, ClassesList, Class, Exam } from "containers";
 //stylesheet
-import './App.css';
+import "./App.css";
 
 library.add(faBars, faTimes, faAngleLeft, faAngleRight);
 
@@ -16,10 +21,10 @@ class App extends Component {
     return (
       <BrowserRouter>
         <div>
-          <Route exact path="/" component={ Login }/>
-          <Route exact path="/classeslist" component={ ClassesList }/>
-          <Route exact path="/class" component={ Class }/>
-          <Route exact path="/exam" component={ Exam } />
+          <Route exact path="/" component={Login} />
+          <Route exact path="/classeslist" component={ClassesList} />
+          <Route exact path="/class" component={Class} />
+          <Route exact path="/exam" component={Exam} />
         </div>
       </BrowserRouter>
     );
