--- src/container/class.css
@@ -1,12 +1,0 @@
-body {
-    padding: 0px;
-    margin: 0;
-    background: linear-gradient(45deg,
-    #EEEEEE 46%, #BEBEBE 0,#BEBEBE 50%, 
-    #EEEEEE 0, #EEEEEE 96%,#BEBEBE 0);
-    background-size: 20px 20px;
-}
-
-.main-body{
-    margin: 20px 200px;
-}
