--- src/container/index.jsx
@@ -1,8 +1,0 @@
-/*
-페이지 컴포넌트들을 파일 하나로 내보냄
-as 이하의 이름으로 내보냄
-*/
-export { default as Login } from 'container/login-container';
-export { default as ClassesList} from 'container/classes-list';
-export { default as Class } from 'container/class';
-export { default as Exam } from 'container/exam'
