+++ src/containers/index.jsx
@@ -1,0 +1,8 @@
+/*
+페이지 컴포넌트들을 파일 하나로 내보냄
+as 이하의 이름으로 내보냄
+*/
+export { default as Login } from "containers/login-container";
+export { default as ClassesList } from "containers/classes-list";
+export { default as Class } from "containers/class";
+export { default as Exam } from "containers/exam";
